增加USB设备通讯命令包打包和拆包对象 增加FC巡更棒协议文档

+++ FCARDIO.Protocol.Door/FC8800/Card/WriteCardList/05CardListBySequence/WriteCardListBySequenceBase.cs
@@ -94,7 +94,7 @@ namespace FCARDIO.Protocol.Door.FC8800.Card.CardListBySequence
                 //未发送完毕，继续发送
                 var buf = GetCmdBuf();
                 WriteCardDetailToBuf(buf);
-                FCPacket.DataLen = (UInt32)buf.ReadableBytes;
+                FCPacket.DataLen = buf.ReadableBytes;
                 CommandReady();//设定命令当前状态为准备就绪，等待发送
             }
         }
